# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auto Tournament API",
    "version": "1.0.0",
    "description": "Run a tournament from a bot or a script: teams, players, brackets, matches and results. The dashboard is a client of this API and nothing more. Which game a tournament runs is a module's business, so the endpoints here are the same whether it is Counter-Strike 2 or a game whose captains report their own scores.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3069",
      "description": "Your MAT instance. Replace with its base URL."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Service token for machine clients (bots, scripts, CI). Send `Authorization: Bearer <token>`, where the token is one configured in API_TOKENS (full admin) or API_TOKENS_READONLY (GET only). Human admins authenticate with a session cookie instead."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Service token for machine clients (bots, scripts, CI). Send `Authorization: Bearer <token>`, where the token is one configured in API_TOKENS (full admin) or API_TOKENS_READONLY (GET only). Human admins authenticate with a session cookie instead."
      },
      "apiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "Service token, as an alternative to `Authorization: Bearer <token>`."
      },
      "atServerToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auto-Tournament-Token",
        "description": "Server-to-API authentication (webhooks, reports, demo uploads, match config fetch). Send `X-Auto-Tournament-Token: <token>`."
      }
    },
...
  },
...
  "paths": {
...
    "/api/events": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "POST /api/events",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "atServerToken": []
          }
        ]
      }
    },
    "/api/events/connections/{matchSlug}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "GET /api/events/connections/:matchSlug",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/events/report": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "POST /api/events/report",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "atServerToken": []
          }
        ]
      }
    },
    "/api/events/server/{serverId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "GET /api/events/server/:serverId",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
...
    "/api/events/{matchSlugOrServerId}": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "POST /api/events/:matchSlugOrServerId",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlugOrServerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "atServerToken": []
          }
        ]
      }
    },
    "/api/events/{matchSlug}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "GET /api/events/:matchSlug",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
...
    "/api/rcon/test": {
      "get": {
        "tags": [
          "RCON"
        ],
        "summary": "Test all RCON connections",
        "description": "Test RCON connectivity to all enabled servers",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Test results"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/rcon/test/{serverId}": {
      "get": {
        "tags": [
          "RCON"
        ],
        "summary": "Test RCON connection to specific server",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection successful"
          },
          "400": {
            "description": "Connection failed"
          }
        }
      }
    },
...
    "/api/settings/igdb/test": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Test the IGDB connection (admin)",
        "description": "Requests a fresh Twitch token and runs a one-row IGDB query with the active credentials.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "`ok: true`, or `ok: false` with an error safe to show"
          }
        }
      }
    },
...
  }
}